docs + moving
